+++ CHANGELOG.rst
@@ -12,3 +12,6 @@ New features
 
 Bug fixes
 ---------
+
+- Fixed bug in output stats that what we called T was really sigma.  Now, it's correctly
+  T = Ixx + Iyy = 2*sigma^2.